(chore) more moonscript into separate C file
